Current working changes: Game ends after 4 turns (2 player turns make up 1 turn, excludes additional turn gained by taking 2 pieces)
Game board has random lava squares, preventing movement and spawning on those squares.

The first team to take 2 of the opponents pieces gets an additional turn, which does not go towards the total amount of game turns.

Known bugs: Blue hen cannot attack or recruit, possibly more.
If you choose a square with nothing on it as the from square, the game will crash.